checkpoint - migrate to using long/Long to represent any integer ID value (rowId, objectId, etc).

+++ SequenceAnalysis/api-src/org/labkey/api/sequenceanalysis/RefNtSequenceModel.java
@@ -62,7 +62,7 @@ public class RefNtSequenceModel implements Serializable
     //deprecated
     private String _sequence;
     private Long _sequenceFile;
-    private Integer _jobId;
+    private Long _jobId;
     private String _category;
     private String _subset;
     private String _locus;
@@ -361,12 +361,12 @@ public class RefNtSequenceModel implements Serializable
         _modified = modified;
     }
 
-    public Integer getJobId()
+    public Long getJobId()
     {
         return _jobId;
     }
 
-    public void setJobId(Integer jobId)
+    public void setJobId(Long jobId)
     {
         _jobId = jobId;
     }
