+++ SequenceAnalysis/api-src/org/labkey/api/sequenceanalysis/SequenceAnalysisService.java
@@ -87,7 +87,7 @@ abstract public class SequenceAnalysisService
 
     abstract public String getUnzippedBaseName(String filename);
 
-    abstract public Integer getExpRunIdForJob(PipelineJob job, boolean throwUnlessFound) throws PipelineJobException;
+    abstract public Long getExpRunIdForJob(PipelineJob job, boolean throwUnlessFound) throws PipelineJobException;
 
     abstract public List<PedigreeRecord> generatePedigree(Collection<String> sampleNames, Container c, User u, DemographicsProvider d);
